requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.0.3
cryptography==2.3.1
decorator==4.3.0
flasgger==0.8.0
Flask==1.0.2
Flask-JWT-Extended==3.12.1
Flask-Testing==0.7.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.3.0
pluggy==0.7.1
psycopg2-binary==2.7.5
py==1.5.4
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.6.4
pylint==1.9.2
pytest==3.7.2
pytest-cov==2.5.1
python-coveralls==2.9.1
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
validators==0.12.2
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
Werkzeug==0.14.1
wrapt==1.10.11
